+++ src/planx/suitability/__init__.py
@@ -17,6 +17,7 @@ from .facility import (
 from .mcda import (
     aras_method,
     copras_method,
+    edas_method,
     electre_i_method,
     electre_iii_method,
     marcos_method,
@@ -50,6 +51,7 @@ __all__ = [
     "electre_iii_method",
     "aras_method",
     "copras_method",
+    "edas_method",
     "marcos_method",
     "waspas_method",
     "mcda_sensitivity_monte_carlo",